+++ CotEditor/Sources/EditorTextView.swift
@@ -1304,12 +1304,19 @@ final class EditorTextView: NSTextView, CurrentLineHighlighting, Themable {
         
         // -> use own dataDetector instead of `checkTextInDocument(_:)` due to performance issue (2018-07)
         let detector = try! NSDataDetector(types: NSTextCheckingResult.CheckingType.link.rawValue)
+        let range = self.string.nsRange
         
-        detector.enumerateMatches(in: self.string, range: self.string.nsRange) { (result, _, _) in
+        self.textStorage?.removeAttribute(.link, range: range)
+        
+        detector.enumerateMatches(in: self.string, range: range) { (result, _, _) in
             guard let result = result, let url = result.url else { return }
             
             self.textStorage?.addAttribute(.link, value: url, range: result.range)
         }
+        
+        // ensure layout to avoid unwanted scroll with cursor move after pasting something
+        // at the latter part of the document. (2018-10 macOS 10.14)
+        self.layoutManager?.ensureLayout(forCharacterRange: range)
     }